Incorporating initialize_project.R script to ensure proper setup
Created initialize_project function to be run at the beginning of the two core R programs (datacleansing.R and timelinecharts.R). This ensure all packages are loaded, and the directory structure is created. Then it downloads the data (from figshare.com) and then the R scripts from Github. This ensure a near-foolproof setup for the R scripts. Also deleted the normalize_street.R function as it is no longer used.